Factor out CameraEncoder, MicrophoneEncoder, and ScreenEncoder
The only structural change is that previously there was a single bit
`destroy: Arc<AtomicBool>` that would stop all encoding when
it was set, but now each Encoder has its own. Previously the
`host.destroy()` method would set that bit, but now it calls a `.stop()`
method for each encoder.

Also: 
 * Moved the transforms from model/mod.rs into model/encode/transform.rs
since they're only used in model/encode.

 * Moved the wrappers from model/mod.rs into model/wrappers.rs just so
that mod.rs isn't a mixed bag of submodule declarations and code definition.
